[JAB] Renamed helpers file to interactionHelpers file
