+++ test/e2e/step-definitions/healthdeclaration-steps.ts
@@ -1,7 +1,7 @@
 import { When, Then, Before } from 'cucumber';
 import { enterPasscode } from './generic-steps';
 import { by } from 'protractor';
-import { getElement, clickElement } from '../../helpers/helpers';
+import { getElement, clickElement } from '../../helpers/interactionHelpers';
 
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
