+++ test/e2e/step-definitions/postdebriefholding-steps.ts
@@ -1,6 +1,6 @@
 import { Then, When, Before } from 'cucumber';
 import { browser, by, ExpectedConditions } from 'protractor';
-import { getElement, clickElement } from '../../helpers/helpers';
+import { getElement, clickElement } from '../../helpers/interactionHelpers';
 
 this.testCategory = 'b';
