+++ test/e2e/step-definitions/testreport-steps.ts
@@ -1,6 +1,6 @@
 import { Then, When, Before } from 'cucumber';
 import { browser, by, element } from 'protractor';
-import { waitForPresenceOfElement, getElement, clickElement } from '../../helpers/helpers';
+import { waitForPresenceOfElement, getElement, clickElement } from '../../helpers/interactionHelpers';
 
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
