+++ test/e2e/step-definitions/waitingroom-steps.ts
@@ -1,7 +1,7 @@
 import { Then, When, Before } from 'cucumber';
 import { enterPasscode } from './generic-steps';
 import { by } from 'protractor';
-import { getElement, clickElement } from '../../helpers/helpers';
+import { getElement, clickElement } from '../../helpers/interactionHelpers';
 
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
