+++ test/helpers/interactionHelpers.ts
